+++ collector/src/test/java/com/bitmovin/analytics/EventDataFactoryTest.kt
@@ -1,6 +1,6 @@
 package com.bitmovin.analytics
 
-import android.test.mock.MockContext
+import android.content.Context
 import com.bitmovin.analytics.data.DeviceInformation
 import com.bitmovin.analytics.data.DeviceInformationProvider
 import com.bitmovin.analytics.data.EventData
@@ -27,7 +27,7 @@ class EventDataFactoryTest {
 
         val myMock = Mockito.mock(DeviceInformationProvider::class.java)
         Mockito.`when`(myMock.getDeviceInformation()).thenReturn(deviceInfo)
-        val mockContext = MockContext()
+        val mockContext = Mockito.mock(Context::class.java)
         val eventData: EventData = EventDataFactory(BitmovinAnalyticsConfig(), mockContext, myMock, userIdProvider).build("impression-id")
 
         assertThat(eventData.deviceInformation.manufacturer).isEqualTo(deviceInfo.manufacturer)
@@ -44,7 +44,7 @@ class EventDataFactoryTest {
 
         val myMock = Mockito.mock(DeviceInformationProvider::class.java)
         Mockito.`when`(myMock.getDeviceInformation()).thenReturn(deviceInfo)
-        val mockContext = MockContext()
+        val mockContext = Mockito.mock(Context::class.java)
         val eventData: EventData = EventDataFactory(BitmovinAnalyticsConfig(), mockContext, myMock, userIdProvider).build("impression-id")
 
         assertThat(eventData.domain).isEqualTo(deviceInfo.packageName)
@@ -57,7 +57,7 @@ class EventDataFactoryTest {
         val deviceInfoMock = Mockito.mock(DeviceInformationProvider::class.java)
         Mockito.`when`(deviceInfoMock.getDeviceInformation()).thenReturn(deviceInfo)
 
-        val mockContext = MockContext()
+        val mockContext = Mockito.mock(Context::class.java)
         val eventData: EventData = EventDataFactory(BitmovinAnalyticsConfig(), mockContext, deviceInfoMock, userIdProvider).build("impression-id")
 
         assertThat(eventData.userId).isEqualTo("my-user-id")
@@ -70,7 +70,7 @@ class EventDataFactoryTest {
         val deviceInfoMock = Mockito.mock(DeviceInformationProvider::class.java)
         Mockito.`when`(deviceInfoMock.getDeviceInformation()).thenReturn(deviceInfo)
 
-        val mockContext = MockContext()
+        val mockContext = Mockito.mock(Context::class.java)
         val impressionId = "impression-id"
         val eventData: EventData = EventDataFactory(BitmovinAnalyticsConfig(), mockContext, deviceInfoMock, userIdProvider).build(impressionId)